Core: Refactor checksum specific methods into common/checksum
