+++ lib/rucio/client/uploadclient.py
@@ -26,6 +26,7 @@ from typing import TYPE_CHECKING, Any, Final, Optional, Union, cast
 from rucio import version
 from rucio.client.client import Client
 from rucio.common.bittorrent import bittorrent_v2_merkle_sha256
+from rucio.common.checksum import GLOBALLY_SUPPORTED_CHECKSUMS, adler32, md5
 from rucio.common.client import detect_client_location
 from rucio.common.config import config_get, config_get_bool, config_get_int
 from rucio.common.constants import RseAttr
@@ -44,7 +45,7 @@ from rucio.common.exception import (
     ScopeNotFound,
     ServiceUnavailable,
 )
-from rucio.common.utils import GLOBALLY_SUPPORTED_CHECKSUMS, adler32, bittorrent_v2_merkle_sha256, execute, generate_uuid, make_valid_did, md5, retry, send_trace
+from rucio.common.utils import execute, generate_uuid, make_valid_did, retry, send_trace
 from rucio.rse import rsemanager as rsemgr
 
 if TYPE_CHECKING:
