+++ lib/rucio/common/utils.py
@@ -17,12 +17,10 @@ import base64
 import datetime
 import errno
 import getpass
-import io
 import ipaddress
 import itertools
 import json
 import logging
-import mmap
 import os
 import os.path
 import re
@@ -32,10 +30,9 @@ import subprocess
 import tempfile
 import threading
 import time
-import zlib
 from collections import OrderedDict
 from enum import Enum
-from functools import partial, wraps
+from functools import wraps
 from io import StringIO
 from itertools import zip_longest
 from typing import TYPE_CHECKING, Any, Optional, TypeVar, Union
@@ -222,137 +219,6 @@ def generate_uuid() -> str:
 
 def generate_uuid_bytes() -> bytes:
     return uuid().bytes
-
-
-# GLOBALLY_SUPPORTED_CHECKSUMS = ['adler32', 'md5', 'sha256', 'crc32']
-GLOBALLY_SUPPORTED_CHECKSUMS = ['adler32', 'md5']
-CHECKSUM_ALGO_DICT = {}
-PREFERRED_CHECKSUM = GLOBALLY_SUPPORTED_CHECKSUMS[0]
-CHECKSUM_KEY = 'supported_checksums'
-
-
-def is_checksum_valid(checksum_name: str) -> bool:
-    """
-    A simple function to check whether a checksum algorithm is supported.
-    Relies on GLOBALLY_SUPPORTED_CHECKSUMS to allow for expandability.
-
-    :param checksum_name: The name of the checksum to be verified.
-    :returns: True if checksum_name is in GLOBALLY_SUPPORTED_CHECKSUMS list, False otherwise.
-    """
-
-    return checksum_name in GLOBALLY_SUPPORTED_CHECKSUMS
-
-
-def set_preferred_checksum(checksum_name: str) -> None:
-    """
-    If the input checksum name is valid,
-    set it as PREFERRED_CHECKSUM.
-
-    :param checksum_name: The name of the checksum to be verified.
-    """
-    if is_checksum_valid(checksum_name):
-        global PREFERRED_CHECKSUM
-        PREFERRED_CHECKSUM = checksum_name
-
-
-def adler32(file: "FileDescriptorOrPath") -> str:
-    """
-    An Adler-32 checksum is obtained by calculating two 16-bit checksums A and B
-    and concatenating their bits into a 32-bit integer. A is the sum of all bytes in the
-    stream plus one, and B is the sum of the individual values of A from each step.
-
-    :param file: file name
-    :returns: Hexified string, padded to 8 values.
-    """
-
-    # adler starting value is _not_ 0
-    adler = 1
-
-    can_mmap = False
-    # try:
-    #    with open(file, 'r+b') as f:
-    #        can_mmap = True
-    # except:
-    #    pass
-
-    try:
-        # use mmap if possible
-        if can_mmap:
-            with open(file, 'r+b') as f:
-                m = mmap.mmap(f.fileno(), 0)
-                # partial block reads at slightly increased buffer sizes
-                for block in iter(partial(m.read, io.DEFAULT_BUFFER_SIZE * 8), b''):
-                    adler = zlib.adler32(block, adler)
-        else:
-            with open(file, 'rb') as f:
-                # partial block reads at slightly increased buffer sizes
-                for block in iter(partial(f.read, io.DEFAULT_BUFFER_SIZE * 8), b''):
-                    adler = zlib.adler32(block, adler)
-
-    except Exception as e:
-        raise Exception('FATAL - could not get Adler-32 checksum of file %s: %s' % (file, e))
-
-    # backflip on 32bit -- can be removed once everything is fully migrated to 64bit
-    if adler < 0:
-        adler = adler + 2 ** 32
-
-    return str('%08x' % adler)
-
-
-CHECKSUM_ALGO_DICT['adler32'] = adler32
-
-
-def md5(file: "FileDescriptorOrPath") -> str:
-    """
-    Runs the MD5 algorithm (RFC-1321) on the binary content of the file named file and returns the hexadecimal digest
-
-    :param file: file name
-    :returns: string of 32 hexadecimal digits
-    """
-    hash_md5 = hashlib.md5()
-    try:
-        with open(file, "rb") as f:
-            list(map(hash_md5.update, iter(lambda: f.read(4096), b"")))
-    except Exception as e:
-        raise Exception('FATAL - could not get MD5 checksum of file %s - %s' % (file, e))
-
-    return hash_md5.hexdigest()
-
-
-CHECKSUM_ALGO_DICT['md5'] = md5
-
-
-def sha256(file: "FileDescriptorOrPath") -> str:
-    """
-    Runs the SHA256 algorithm on the binary content of the file named file and returns the hexadecimal digest
-
-    :param file: file name
-    :returns: string of 32 hexadecimal digits
-    """
-    with open(file, "rb") as f:
-        bytes_ = f.read()  # read entire file as bytes
-        readable_hash = hashlib.sha256(bytes_).hexdigest()
-        print(readable_hash)
-        return readable_hash
-
-
-CHECKSUM_ALGO_DICT['sha256'] = sha256
-
-
-def crc32(file: "FileDescriptorOrPath") -> str:
-    """
-    Runs the CRC32 algorithm on the binary content of the file named file and returns the hexadecimal digest
-
-    :param file: file name
-    :returns: string of 32 hexadecimal digits
-    """
-    prev = 0
-    for eachLine in open(file, "rb"):
-        prev = zlib.crc32(eachLine, prev)
-    return "%X" % (prev & 0xFFFFFFFF)
-
-
-CHECKSUM_ALGO_DICT['crc32'] = crc32
 
 
 def str_to_date(string: str) -> Optional[datetime.datetime]:
