+++ lib/rucio/rse/protocols/rclone.py
@@ -17,8 +17,9 @@ import logging
 import os
 
 from rucio.common import exception
+from rucio.common.checksum import PREFERRED_CHECKSUM
 from rucio.common.config import get_config_dirs
-from rucio.common.utils import PREFERRED_CHECKSUM, execute
+from rucio.common.utils import execute
 from rucio.rse.protocols import protocol
