+++ lib/rucio/rse/protocols/xrootd.py
@@ -16,7 +16,8 @@ import logging
 import os
 
 from rucio.common import exception
-from rucio.common.utils import PREFERRED_CHECKSUM, execute
+from rucio.common.checksum import PREFERRED_CHECKSUM
+from rucio.common.utils import execute
 from rucio.rse.protocols import protocol
