+++ tests/test_rse_protocol_rclone.py
@@ -16,7 +16,8 @@ import os
 
 import pytest
 
-from rucio.common.utils import PREFERRED_CHECKSUM, execute, set_preferred_checksum
+from rucio.common.checksum import PREFERRED_CHECKSUM, set_preferred_checksum
+from rucio.common.utils import execute
 from rucio.rse import rsemanager
 from rucio.tests.common import load_test_conf_file, skip_rse_tests_with_accounts
