+++ tests/test_upload.py
@@ -23,10 +23,11 @@ import pytest
 
 from rucio.client.client import Client
 from rucio.client.uploadclient import UploadClient
+from rucio.common.checksum import adler32
 from rucio.common.config import config_add_section, config_set
 from rucio.common.constants import RseAttr
 from rucio.common.exception import InputValidationError, NoFilesUploaded, NotAllFilesUploaded, ResourceTemporaryUnavailable
-from rucio.common.utils import adler32, generate_uuid
+from rucio.common.utils import generate_uuid
 from rucio.core.rse import add_protocol, add_rse_attribute
 
 
@@ -406,13 +407,17 @@ def upload_client_registration_fail():
     logger = logging.getLogger('upload_client')
     logger.addHandler(logging.StreamHandler())
     logger.setLevel(logging.DEBUG)
+
     # modify the client object used by upload_client so that replica registration fails
     class RegistrationFailureClient(Client):
+
         def __init__(self, **args):
             super(RegistrationFailureClient, self).__init__(**args)
+
         def update_replicas_states(self, rse, files):
             # simulate server timing out
             raise ResourceTemporaryUnavailable
+
     return UploadClient(logger=logger, _client=RegistrationFailureClient())
